# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: User not found in context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Requested file not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: Internal server error (S3, DB, or parse failure)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
paths:
...
  /ro_risk_dist:
    get:
      summary: Get risk distribution counts across all regional offices
      operationId: GetRORiskDist
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Keyed by RO name (Bangalore, Chandigarh, Delhi, Guwahati, Hyderabad, Lucknow, Mumbai, Ranchi)
                    additionalProperties:
                      type: object
                      properties:
                        high_risk_count: { type: integer }
                        med_risk_count: { type: integer }
                        low_risk_count: { type: integer }
                        no_risk_count: { type: integer }
                  file: { type: string }
                  regional_office: { type: string }
                  requested_by: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /geteaandreg:
    get:
      summary: Get a flat list of all EAs or registrars for caller's RO
      operationId: GetEaAndReg
      parameters:
        - name: value
          in: query
          required: true
          schema: { type: string, enum: [ea, reg] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { type: string } }
                  count: { type: integer }
                  type: { type: string }
                  regional_office: { type: string }
        "400":
          description: Missing or invalid 'value' parameter
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  /selected_eas:
    post:
      summary: Get risk distribution for a chosen set of EAs
      operationId: GetSelectedEAs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selected_eas]
              properties:
                selected_eas:
                  type: array
                  items: { type: string }
                  minItems: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ea_distribution:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        high_risk: { type: integer }
                        med_risk: { type: integer }
                        low_risk: { type: integer }
                        no_risk: { type: integer }
        "400":
          description: Invalid body or empty selected_eas
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /all_eas:
    get:
      summary: Get all EA names for caller's RO (sorted)
      operationId: GetAllEAs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  eas: { type: array, items: { type: string } }
                  count: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /report_anomaly:
    post:
      summary: Report an anomaly for a packet
      operationId: ReportAnomaly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eid, opt_id]
              properties:
                eid: { type: string }
                opt_id: { type: string }
                anomaly_category: { type: string }
                anomaly_type:
                  type: array
                  description: Only the first element is used.
                  items:
                    type: object
                    properties:
                      anomaly_category: { type: string }
                      anomaly_code: { type: string }
                      anomaly_name: { type: string }
                      reason:
                        type: object
                        properties:
                          error_category: { type: string }
                date_created: { type: string }
                enrolnment_type: { type: string }
                opt_state: { type: string }
                opt_district: { type: string }
                pkt_source: { type: string }
                pkt_updt_type:
                  type: array
                  items: { type: string }
                  description: Joined with ", " before insert.
                remarks:
                  type: array
                  items: { type: string }
                  description: Joined with ", " before insert.
                station_machine_code: { type: string }
                station_no: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  eid: { type: string }
        "400":
          description: Invalid JSON or missing eid/opt_id
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500":
          description: Failed to insert anomaly record
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }